backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
firebase-admin==6.3.0
PyJWT==2.8.0
user-agents==2.2.0